Add pipeline run diagram preview
Fixes #77
